// models/user/userModel.ts
import { GroupEnum } from '../../enums/group.emum';

export interface UserModel {
  email?: string;
  password?: string;
  acceptNewStudents?: boolean;
  acceptsSubstitutions?: boolean;
  additionalName1?: string;
  additionalName2?: string;
  additionalPhone1?: string;
  additionalPhone2?: string;
  biography?: string;
  biographyPl?: string;
  biographyRu?: string;
  biographyTr?: string;
  churnComment?: string;
  churnDateManual?: string;
  churnType?: string;
  classCost?: string;
  comment?: string;
  education?: string;
  gender?: string;
  grafanaTeams?: string;
  group?: GroupEnum;
  isTest?: string;
  isVerifiedChurn?: boolean;
  lang?: string;
  name?: string;
  paypalEmail?: string;
  paypalPayerId?: string;
  phone?: string;
  photoFileId?: string;
  regionCode?: string;
  sendCall?: string;
  sendSms?: string;
  teacherIsMentor?: boolean;
  teacherIsNearNative?: boolean;
  teacherL0Certified?: boolean;
  teacherMentorId?: string;
  teacherVideoUrl?: string;
  timezone?: string;
  trialRank?: number;
  utmData?: string;
  id?: number;
  countryOfOrigin?: string;
  balance?: number;
  balanceInNativeLessons?: number;
  balanceInNearNativeLessons?: number;
  studentId?: number;
  studentId2?: number;
}